requirements.txt
hydra-core~=1.1.0
omegaconf~=2.1.0
pytorch-lightning~=1.3.8
pytorch-lightning-bolts
transformers~=4.0.0
torch~=1.7.0
torchmetrics~=0.4.1